pyproject.toml
[tool.poetry]
name = "timeboard"
version = "0.1.0"
description = ""
authors = ["Kazem <musakazem440@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2"
django-admin-extra-buttons = "^2.1"
matplotlib = "^3.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
